Library information

Constants

MPI.MPI_VERSION — Constant

MPI_VERSION :: VersionNumber

The supported version of the MPI standard.

External links

MPI_Get_version man page: OpenMPI, MPICH

MPI.MPI_LIBRARY — Constant

MPI_LIBRARY :: MPIImpl

The current MPI implementation: this is determined by

See also

MPIImpl

MPI.MPI_LIBRARY_VERSION — Constant

MPI_LIBRARY_VERSION :: VersionNumber

The version of the MPI library

MPI.MPI_LIBRARY_VERSION_STRING — Constant

MPI_LIBRARY_VERSION_STRING :: String

The full version string provided by the library

External links

MPI_Get_library_version man page: OpenMPI, MPICH

Enums

MPI.MPIImpl — Type

MPIImpl

An enum corresponding to known MPI implementations

UnknownMPI: unable to determine MPI implementation

MPICH: MPICH

OpenMPI: Open MPI

MicrosoftMPI: Microsoft MPI

IntelMPI: Intel MPI

SpectrimMPI: IBM Spectrum MPI

MVAPICH: MVAPICH

CrayMPICH: Part of the Cray Message Passing Toolkit (MPT)

See also

MPI_LIBRARY

Functions

MPI.has_cuda — Function

MPI.has_cuda()

Check if the MPI implementation is known to have CUDA support. Currently only Open MPI provides a mechanism to check, so it will return false with other implementations (unless overriden).

This can be overriden by setting the JULIA_MPI_HAS_CUDA environment variable to true or false.

MPI.identify_implementation — Function

impl, version = identify_implementation()

Attempt to identify the MPI implementation based on MPI_LIBRARY_VERSION_STRING. Returns a triple of values:

impl: a value of type MPIImpl

version: a VersionNumber of the library, or nothing if it cannot be determined.

This function is only intended for internal use. Users should use MPI_LIBRARY, MPI_LIBRARY_VERSION.