MPIPreferences.jl

Consts

MPIPreferences.binary — Constant

binary :: String

The currently selected binary.

MPIPreferences.abi — Constant

abi :: String

The ABI of the currently selected binary.

Changing implementations

MPIPreferences.use_system_binary — Function

use_system_binary(;library_names, mpiexec="mpiexec", abi=nothing, export_prefs=false, force=true)

Switches the underlying MPI implementation to a system provided one. We will search for a matching library with the given library_names. You can provide a path to mpiexec if the system provides a different one, or if it is not in the system path.

The abi of the found library is determined automatically using identify_abi.

The export_prefs option determines whether the preferences being set should be stored within LocalPreferences.toml or Project.toml.

MPIPreferences.use_jll_binary — Function

use_jll_binary(binary; export_prefs=false, force=true)

Switches the underlying MPI implementation to one provided by JLL packages. Available options are:

MicrosoftMPI_jll (Only option and default on Winddows)

MPICH_jll (Default on every other platform)

OpenMPI_jll

MPItrampoline_jll

The export_prefs option determines whether the preferences being set should be stored within LocalPreferences.toml or Project.toml.

Utils

MPIPreferences.identify_abi — Function

identify_abi(libmpi)

Identify the MPI implementation from the library version string