Groups

An MPI group is a set of process identifiers identified by their rank (see MPI.Comm_rank and MPI.Group_rank). They are used within a communicator to describe the participants in a communication universe.

Types and enums

MPI.Group — Type

MPI.Group

An MPI Group object.

MPI.Comparison — Type

Comparison

An enum denoting the result of Comm_compare:

MPI.IDENT: the objects are handles for the same object (identical groups and same contexts).

MPI.CONGRUENT: the underlying groups are identical in constituents and rank order; these communicators differ only by context.

MPI.SIMILAR: members of both objects are the same but the rank order differs.

MPI.UNEQUAL: otherwise

Functions

Operations

MPI.Group_size — Function

Group_size(group::Group)

The number of processes involved in group.

External links

MPI_Group_size man page: OpenMPI, MPICH

MPI.Group_rank — Function

Group_rank(group::Group)

The rank of the process in the particular group.

Returns an integer in the range 0:MPI.Group_size()-1.

External links

MPI_Group_rank man page: OpenMPI, MPICH