Topology

MPI.Cart_coords — Function

Cart_coords(comm::Comm)

Determine coordinates of local process in Cartesian communicator.

See also Cart_coords!.

MPI.Cart_coords! — Function

Cart_coords!(comm::Comm, rank::Integer, coords)

Determine coordinates of a given process in Cartesian communicator.

External links

MPI_Cart_coords man page: OpenMPI, MPICH

MPI.Cart_create — Function

comm_cart = Cart_create(comm_old::Comm, [ndims::Integer], dims, periods, reorder)

Create new MPI communicator with Cartesian structure from an existent communicator.

External links

MPI_Cart_create man page: OpenMPI, MPICH

MPI.Cart_get — Function

dims, periods, coords = Cart_get(comm::Comm)

Obtain information on the Cartesian topology of dimension N underlying the communicator comm. This is specified by two Cint arrays of N elements for the number of processes and periodicity properties along each Cartesian dimension. A third Cint array is returned, containing the Cartesian coordinates of the calling process.

External links

MPI_Cart_get man page: OpenMPI, MPICH

MPI.Cart_rank — Function

rank = Cart_rank(comm::Comm, coords)

Determine process rank in communicator comm with Cartesian structure. The coords array specifies the Cartesian coordinates of the process.

External links

MPI_Cart_rank man page: OpenMPI, MPICH

MPI.Cart_shift — Function

rank_source, rank_dest = Cart_shift(comm::Comm, direction::Integer, disp::Integer)

Return the source and destination ranks associated to a shift along a given direction.

External links

MPI_Cart_shift man page: OpenMPI, MPICH

MPI.Cart_sub — Function

comm_sub = Cart_sub(comm::Comm, remain_dims)

Create lower-dimensional Cartesian communicator from existent Cartesian topology.

remain_dims should be a boolean vector specifying the dimensions that should be kept in the generated subgrid.

External links

MPI_Cart_sub man page: OpenMPI, MPICH

MPI.Cartdim_get — Function

ndims = Cartdim_get(comm::Comm)

Return number of dimensions of the Cartesian topology associated with the communicator comm.

External links

MPI_Cartdim_get man page: OpenMPI, MPICH

MPI.Dims_create! — Function

Dims_create!(nnodes::Integer, [ndims::Integer], dims)

Create a division of nnodes processes in a Cartesian grid.

External links

MPI_Dims_create man page: OpenMPI, MPICH