Advanced

Object handling

MPI.free — Function

MPI.free(obj)

Free the MPI object handle obj. This is typically used as the finalizer, and so need not be called directly unless otherwise noted.

MPI.refcount_inc — Function

refcount_inc()

Increment the MPI reference counter. This should be called at initialization of any object which calls an MPI routine in its finalizer. A matching refcount_dec should be added to the finalizer.

For more details, see Finalizers.

MPI.refcount_dec — Function

refcount_dec()

Decrement the MPI reference counter. This should be added after an MPI call in an object finalizer, with a matching refcount_inc when the object is initialized.

For more details, see Finalizers.

Datatype objects

MPI.mpitype — Function

mpitype(T)

Returns the MPI Datatype code for a given type T. In the case the the type does not exist, it is created and then returned. The dictonary is defined in __init__ so the module can be precompiled

MPI.Type_Create_Subarray — Function

Type_Create_Subarray(ndims::Integer, array_of_sizes::MPIBuffertype{Cint},
                     array_of_subsizes::MPIBuffertype{Cint},
                     array_of_starts::MPIBuffertype{Cint}, order::Integer, oldtype)

Creates a data type describing an ndims-dimensional subarray of size array_of_subsizes of an ndims-dimensional array of size array_of_sizes and element type oldtype, starting at the top-left location array_of_starts. Zero-based indexing is assumed. The parameter order refers to the memory layout of the parent array, and can be either MPI_ORDER_C or MPI_ORDER_FORTRAN. Note that, like other MPI data types, the type returned by this function should be committed with MPI_Type_commit.

Operator objects

MPI.Op — Type

Op

An MPI reduction operator, for use with Reduce/Scan collective operations to wrap binary operators. MPI.jl will perform this conversion automatically.

Usage

Op(op, T=Any; iscommutative=false)

Wrap the Julia reduction function op for arguments of type T. op is assumed to be associative, and if iscommutative is true, assumed to be commutative as well.

See also

Reduce!/Reduce

Allreduce!/Allreduce

Scan!/Scan

Exscan!/Exscan

Info objects

MPI.Info — Type

Info <: AbstractDict{Symbol,String}

MPI.Info objects store key-value pairs, and are typically used for passing optional arguments to MPI functions.

Usage

These will typically be hidden from user-facing APIs by splatting keywords, e.g.

function f(args...; kwargs...)
    info = Info(kwargs...)
    # pass `info` object to `ccall`
end

For manual usage, Info objects act like Julia Dict objects:

info = Info(init=true) # keyword argument is required
info[key] = value
x = info[key]
delete!(info, key)

If init=false is used in the costructor (the default), a "null" Info object will be returned: no keys can be added to such an object.

MPI.infoval — Function

infoval(x)

Convert Julia object x to a string representation for storing in an Info object.

The MPI specification allows passing strings, Boolean values, integers, and lists.