Communicators

An MPI communicator specifies the communication context for a communication operation. In particular, it specifies the set of processes which share the context, and assigns each each process a unique rank (see MPI.Comm_rank) taking an integer value in 0:n-1, where n is the number of processes in the communicator (see MPI.Comm_size.

Types and enums

MPI.Comm — Type

MPI.Comm

An MPI Communicator object.

MPI.Comparison — Type

Comparison

An enum denoting the result of Comm_compare:

MPI.IDENT: the objects are handles for the same object (identical groups and same contexts).

MPI.CONGRUENT: the underlying groups are identical in constituents and rank order; these communicators differ only by context.

MPI.SIMILAR: members of both objects are the same but the rank order differs.

MPI.UNEQUAL: otherwise

Constants

MPI.COMM_WORLD — Constant

MPI.COMM_WORLD

A communicator containing all processes with which the local rank can communicate at initialization. In a typical "static-process" model, this will be all processes.

MPI.COMM_SELF — Constant

MPI.COMM_SELF

A communicator containing only the local process.

Functions

Operations

MPI.Comm_size — Function

Comm_size(comm::Comm)

The number of processes involved in communicator.

See also

MPI.Comm_rank.

External links

MPI_Comm_size man page: OpenMPI, MPICH

MPI.Comm_rank — Function

Comm_rank(comm::Comm)

The rank of the process in the particular communicator's group.

Returns an integer in the range 0:MPI.Comm_size()-1.

See also

MPI.Comm_size.

External links

MPI_Comm_rank man page: OpenMPI, MPICH

MPI.Comm_compare — Function

Comm_compare(comm1::Comm, comm2::Comm)::MPI.Comparison

Compare two communicators, returning an element of the Comparison enum.

External links

MPI_Comm_compare man page: OpenMPI, MPICH

Constructors

MPI.Comm_dup — Function

Comm_dup(comm::Comm)

External links

MPI_Comm_dup man page: OpenMPI, MPICH

MPI.Comm_get_parent — Function

Comm_get_parent()

External links

MPI_Comm_get_parent man page: OpenMPI, MPICH

MPI.Comm_spawn — Function

Comm_spawn(command, argv::Vector{String}, nprocs::Integer, comm::Comm[, errors::Vector{Cint}]; kwargs...)

External links

MPI_Comm_spawn man page: OpenMPI, MPICH

MPI.Comm_split — Function

Comm_split(comm::Comm, color::Integer, key::Integer)

External links

MPI_Comm_split man page: OpenMPI, MPICH

MPI.Comm_split_type — Function

Comm_split_type(comm::Comm, split_type::Integer, key::Integer; kwargs...)

External links

MPI_Comm_split_type man page: OpenMPI, MPICH

MPI.Intercomm_merge — Function

Intercomm_merge(intercomm::Comm, flag::Bool)

External links

MPI_Intercomm_merge man page: OpenMPI, MPICH

Miscellaneous

MPI.universe_size