Environment

Functions

MPI.Abort — Function

Abort(comm::Comm, errcode::Integer)

Make a “best attempt” to abort all tasks in the group of comm. This function does not require that the invoking environment take any action with the error code. However, a Unix or POSIX environment should handle this as a return errorcode from the main program.

External links

MPI_Abort man page: OpenMPI, MPICH

MPI.Init — Function

Init()

Initialize MPI in the current process.

All MPI programs must contain exactly one call to MPI.Init(). In particular, note that it is not valid to call MPI.Init again after calling MPI.Finalize.

The only MPI functions that may be called before MPI.Init() are MPI.Initialized and MPI.Finalized.

External links

MPI_Init man page: OpenMPI, MPICH

MPI.Initialized — Function

Initialized()

Returns true if MPI.Init has been called, false otherwise.

It is unaffected by MPI.Finalize, and is one of the few functions that may be called before MPI.Init.

External links

MPI_Intialized man page: OpenMPI, MPICH

MPI.Finalize — Function

Finalize()

Marks MPI state for cleanup. This should be called after Init, at most once, and no further MPI calls (other than Initialized or Finalized) should be made after it is called.

Note that this does not correspond exactly to MPI_FINALIZE in the MPI specification. In particular:

It may not finalize MPI immediately. Julia will wait until all MPI-related objects are garbage collected before finalizing MPI. As a result, Finalized() may return false after Finalize() has been called. See Finalizers for more details.

It is optional: Init will automatically insert a hook to finalize MPI when Julia exits.

External links

MPI_Finalize man page: OpenMPI, MPICH

MPI.Finalized — Function

Finalized()

Returns true if MPI.Finalize has completed, false otherwise.

It is safe to call before MPI.Init and after MPI.Finalize.

External links

MPI_Finalized man page: OpenMPI, MPICH

MPI.refcount_inc — Function

refcount_inc()

Increment the MPI reference counter. This should be called at initialization of any object which calls an MPI routine in its finalizer. A matching refcount_dec should be added to the finalizer.

For more details, see Finalizers.

MPI.refcount_dec — Function

refcount_dec()

Decrement the MPI reference counter. This should be added after an MPI call in an object finalizer, with a matching refcount_inc when the object is initialized.

For more details, see Finalizers.