Function reference

The following functions are currently wrapped, with the convention: MPI_Fun => MPI.Fun

Constants like MPI_SUM are wrapped as MPI.SUM. Note also that arbitrary Julia functions f(x,y) can be passed as reduction operations to the MPI Allreduce and Reduce functions.

Datatype functions

Julia Function (assuming import MPI)	C Function

MPI.Get_address	MPI_Get_address

MPI.mpitype	MPI_Type_create_struct/MPI_Type_commit

Note

mpitype is not strictly a wrapper for MPI_Type_create_struct and MPI_Type_commit, it also is an accessor for previously created types.

Missing docstring.

Missing docstring for MPI.Get_address. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.mpitype. Check Documenter's build log for details.

One-sided communication

Julia Function (assuming import MPI)	C Function

MPI.Win_create	MPI_Win_create

MPI.Win_create_dynamic	MPI_Win_create_dynamic

MPI.Win_allocate_shared	MPI_Win_allocate_shared

MPI.Win_shared_query	MPI_Win_shared_query

MPI.Win_attach	MPI_Win_attach

MPI.Win_detach	MPI_Win_detach

MPI.Win_fence	MPI_Win_fence

MPI.Win_flush	MPI_Win_flush

MPI.Win_free	MPI_Win_free

MPI.Win_sync	MPI_Win_sync

MPI.Win_lock	MPI_Win_lock

MPI.Win_unlock	MPI_Win_unlock

MPI.Get	MPI_Get

MPI.Put	MPI_Put

MPI.Fetch_and_op	MPI_Fetch_and_op

MPI.Accumulate	MPI_Accumulate

MPI.Get_accumulate	MPI_Get_accumulate

MPI.Win_create — Function

MPI.Win_create(base::Array, comm::Comm; infokws...)

Create a window over the array base, returning a Win object used by these processes to perform RMA operations

This is a collective call over comm.

infokws are info keys providing optimization hints.

MPI.free should be called on the Win object once operations have been completed.

MPI.Win_create_dynamic — Function

MPI.Win_create_dynamic(comm::Comm; infokws...)

Create a dynamic window returning a Win object used by these processes to perform RMA operations

This is a collective call over comm.

infokws are info keys providing optimization hints.

MPI.free should be called on the Win object once operations have been completed.

MPI.Win_allocate_shared — Function

(win, ptr) = MPI.Win_allocate_shared(T, len, comm::Comm; infokws...)

Create and allocate a shared memory window for objects of type T of length len, returning a Win and a Ptr{T} object used by these processes to perform RMA operations

This is a collective call over comm.

infokws are info keys providing optimization hints.

MPI.free should be called on the Win object once operations have been completed.

Missing docstring.

Missing docstring for MPI.Win_shared_query. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_attach. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_detach. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_fence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_flush. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_free. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_sync. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_lock. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_unlock. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Get. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Put. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Fetch_and_op. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Accumulate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Get_accumulate. Check Documenter's build log for details.

Info objects

MPI.Info — Type

Info <: AbstractDict{Symbol,String}

MPI.Info objects store key-value pairs, and are typically used for passing optional arguments to MPI functions.

Usage

These will typically be hidden from user-facing APIs by splatting keywords, e.g.

function f(args...; kwargs...)
    info = Info(kwargs...)
    # pass `info` object to `ccall`
end

For manual usage, Info objects act like Julia Dict objects:

info = Info(init=true) # keyword argument is required
info[key] = value
x = info[key]
delete!(info, key)

If init=false is used in the costructor (the default), a "null" Info object will be returned: no keys can be added to such an object.

MPI.infoval — Function

infoval(x)

Convert Julia object x to a string representation for storing in an Info object.

The MPI specification allows passing strings, Boolean values, integers, and lists.