Function reference

Function reference

The following functions are currently wrapped, with the convention: MPI_Fun => MPI.Fun

Constants like MPI_SUM are wrapped as MPI.SUM. Note also that arbitrary Julia functions f(x,y) can be passed as reduction operations to the MPI Allreduce and Reduce functions.

Communicators

Julia interfaces to the Fortran versions of the MPI functions. Since the C and Fortran communicators are different, if a C communicator is required (e.g., to interface with a C library), this can be achieved with the Fortran to C communicator conversion:

juliacomm = MPI.COMM_WORLD
ccomm = MPI.CComm(juliacomm)

Administrative functions

Julia Function (assuming import MPI)	Fortran Function

MPI.Abort	MPI_Abort

MPI.Comm_dup	MPI_Comm_dup

MPI.Comm_free	MPI_Comm_free

MPI.Comm_get_parent	MPI_Comm_get_parent

MPI.Comm_rank	MPI_Comm_rank

MPI.Comm_size	MPI_Comm_size

MPI.Comm_spawn	MPI_Comm_spawn

MPI.Finalize	MPI_Finalize

MPI.Finalized	MPI_Finalized

MPI.Get_address	MPI_Get_address

MPI.Init	MPI_Init

MPI.Initialized	MPI_Initialized

MPI.Intercomm_merge	MPI_Intercomm_merge

MPI.mpitype	MPI_Type_create_struct/MPI_Type_commit

Note

mpitype is not strictly a wrapper for MPI_Type_create_struct and MPI_Type_commit, it also is an accessor for previously created types.

MPI.Abort — Function.

Abort(comm::Comm, errcode::Integer)

Make a “best attempt” to abort all tasks in the group of comm. This function does not require that the invoking environment take any action with the error code. However, a Unix or POSIX environment should handle this as a return errorcode from the main program.

source

Missing docstring.

Missing docstring for MPI.Comm_dup. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Comm_free. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Comm_get_parent. Check Documenter's build log for details.

MPI.Comm_rank — Function.

Comm_rank(comm:Comm)

The rank of the process in the particular communicator’s group.

Returns an integer in the range 0:MPI.Comm_size()-1.

source

MPI.Comm_size — Function.

Comm_size(comm:Comm)

The number of processes involved in communicator.

source

Missing docstring.

Missing docstring for MPI.Comm_spawn. Check Documenter's build log for details.

MPI.Finalize — Function.

Finalize()

Marks MPI state for cleanup. This should be called after Init, at most once, and no further MPI calls (other than Initialized or Finalized) should be made after it is called.

Note that this does not correspond exactly to MPI_FINALIZE in the MPI specification. In particular:

It may not finalize MPI immediately. Julia will wait until all MPI-related objects are garbage collected before finalizing MPI. As a result, Finalized() may return false after Finalize() has been called. See Finalizers for more details.

It is optional: Init will automatically insert a hook to finalize MPI when Julia exits.

source

MPI.Finalized — Function.

Finalized()

Returns true if MPI.Finalize has completed, false otherwise.

It is safe to call before MPI.Init and after MPI.Finalize.

source

Missing docstring.

Missing docstring for MPI.Get_address. Check Documenter's build log for details.

MPI.Init — Function.

Init()

Initialize MPI in the current process.

All MPI programs must contain exactly one call to MPI.Init().

The only MPI functions that may be called before MPI.Init() are MPI.Initialized and MPI.Finalized.

source

MPI.Initialized — Function.

Initialized()

Returns true if MPI.Init has been called, false otherwise.

It is unaffected by MPI.Finalize, and is one of the few functions that may be called before MPI.Init.

source

Missing docstring.

Missing docstring for MPI.Intercomm_merge. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.mpitype. Check Documenter's build log for details.

MPI.refcount_inc — Method.

refcount_inc()

Increment the MPI reference counter. This should be called at initialization of any object which calls an MPI routine in its finalizer. A matching refcount_dec should be added to the finalizer.

For more details, see Finalizers.

source

MPI.refcount_dec — Method.

refcount_dec()

Decrement the MPI reference counter. This should be added after an MPI call in an object finalizer, with a matching refcount_inc when the object is initialized.

For more details, see Finalizers.

source

Point-to-point communication

Julia Function (assuming import MPI)	Fortran Function

MPI.Cancel!	MPI_Cancel

MPI.Get_count	MPI_Get_count

MPI.Iprobe	MPI_Iprobe

MPI.Irecv!	MPI_Irecv

MPI.Isend	MPI_Isend

MPI.Probe	MPI_Probe

MPI.Recv!	MPI_Recv

MPI.Send	MPI_Send

MPI.Test!	MPI_Test

MPI.Testall!	MPI_Testall

MPI.Testany!	MPI_Testany

MPI.Testsome!	MPI_Testsome

MPI.Wait!	MPI_Wait

MPI.Waitall!	MPI_Waitall

MPI.Waitany!	MPI_Waitany

MPI.Waitsome!	MPI_Waitsome

Missing docstring.

Missing docstring for MPI.Cancel!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Get_count. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Iprobe. Check Documenter's build log for details.

MPI.Irecv! — Function.

Irecv!(buf::MPIBuffertype{T}, count::Integer, src::Integer, tag::Integer,
       comm::Comm) where T

Starts a nonblocking receive of up to count elements into buf from MPI rank src of communicator comm using with the message tag tag

Returns the communication Request for the nonblocking receive.

source

Irecv!(buf::Array{T}, src::Integer, tag::Integer, comm::Comm) where T

Starts a nonblocking receive into buf from MPI rank src of communicator comm using with the message tag tag

Returns the communication Request for the nonblocking receive.

source

Irecv!(buf::SubArray{T}, src::Integer, tag::Integer, comm::Comm) where T

Starts a nonblocking receive into SubArray buf from MPI rank src of communicator comm using with the message tag tag. Note that buf must be contiguous.

Returns the communication Request for the nonblocking receive.

source

MPI.Isend — Function.

Isend(buf::MPIBuffertype{T}, count::Integer, dest::Integer, tag::Integer,
      comm::Comm) where T

Starts a nonblocking send of count elements of buf to MPI rank dest of communicator comm using with the message tag tag

Returns the commication Request for the nonblocking send.

source

Isend(buf::Array{T}, dest::Integer, tag::Integer, comm::Comm) where T

Starts a nonblocking send of buf to MPI rank dest of communicator comm using with the message tag tag

Returns the commication Request for the nonblocking send.

source

Isend(buf::SubArray{T}, dest::Integer, tag::Integer, comm::Comm) where T

Starts a nonblocking send of SubArray buf to MPI rank dest of communicator comm using with the message tag tag. Note that the buf must be contiguous.

Returns the commication Request for the nonblocking send.

source

Isend(obj::T, dest::Integer, tag::Integer, comm::Comm) where T

Starts a nonblocking send of obj to MPI rank dest of communicator comm using with the message tag tag.

Returns the commication Request for the nonblocking send.

source

Missing docstring.

Missing docstring for MPI.Probe. Check Documenter's build log for details.

MPI.Recv! — Function.

Recv!(buf::MPIBuffertype{T}, count::Integer, src::Integer, tag::Integer,
      comm::Comm) where T

Completes a blocking receive of up to count elements into buf from MPI rank src of communicator comm using with the message tag tag

Returns the Status of the receive

source

Recv!(buf::Array{T}, src::Integer, tag::Integer, comm::Comm) where T

Completes a blocking receive into buf from MPI rank src of communicator comm using with the message tag tag

Returns the Status of the receive

source

Recv!(buf::SubArray{T}, src::Integer, tag::Integer, comm::Comm) where T

Completes a blocking receive into SubArray buf from MPI rank src of communicator comm using with the message tag tag. Note that buf must be contiguous.

Returns the Status of the receive

source

MPI.Send — Function.

Send(buf::MPIBuffertype{T}, count::Integer, dest::Integer, tag::Integer,
     comm::Comm) where T

Complete a blocking send of count elements of buf to MPI rank dest of communicator comm using with the message tag tag

source

Send(buf::Array{T}, dest::Integer, tag::Integer, comm::Comm) where T

Complete a blocking send of buf to MPI rank dest of communicator comm using with the message tag tag

source

Send(buf::SubArray{T}, dest::Integer, tag::Integer, comm::Comm) where T

Complete a blocking send of SubArray buf to MPI rank dest of communicator comm using with the message tag tag. Note that the buf must be contiguous.

source

Send(obj::T, dest::Integer, tag::Integer, comm::Comm) where T

Complete s blocking send of obj to MPI rank dest of communicator comm using with the message tag tag.

source

Missing docstring.

Missing docstring for MPI.Test!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Testall!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Testany!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Testsome!. Check Documenter's build log for details.

MPI.Wait! — Function.

Wait!(req::Request)

Wait on the request req to be complete. Returns the Status of the request.

source

MPI.Waitall! — Function.

Waitall!(reqs::Array{Request,1})

Wait on all the requests in the array reqs to be complete. Returns an arrays of the all the requests statuses.

source

MPI.Waitany! — Function.

Waitany!(reqs::Array{Request,1})

Wait on any the requests in the array reqs to be complete. Returns the index of the completed request and its Status as a tuple.

source

Missing docstring.

Missing docstring for MPI.Waitsome!. Check Documenter's build log for details.

Collective communication

Non-Allocating Julia Function	Allocating Julia Function	Fortran Function	Supports MPI_IN_PLACE

MPI.Allgather!	MPI.Allgather	MPI_Allgather	✅

MPI.Allgatherv!	MPI.Allgatherv	MPI_Allgatherv	✅

MPI.Allreduce!	MPI.Allreduce	MPI_Allreduce	✅

MPI.Alltoall!	MPI.Alltoall	MPI_Alltoall	✅

MPI.Alltoallv!	MPI.Alltoallv	MPI_Alltoallv	❌

–	MPI.Barrier	MPI_Barrier	❌

MPI.Bcast!	MPI.Bcast!	MPI_Bcast	❌

–	MPI.Exscan	MPI_Exscan	❌

MPI.Gather!	MPI.Gather	MPI_Gather	Gather_in_place!

MPI.Gatherv!	MPI.Gatherv	MPI_Gatherv	Gatherv_in_place!

MPI.Reduce!	MPI.Reduce	MPI_Reduce	Reduce_in_place!

MPI.Scan	MPI.Scan	MPI_Scan	missing

MPI.Scatter!	MPI.Scatter	MPI_Scatter	Scatter_in_place!

MPI.Scatterv!	MPI.Scatterv	MPI_Scatterv	Scatterv_in_place!

The non-allocating Julia functions map directly to the corresponding MPI operations, after asserting that the size of the output buffer is sufficient to store the result.

The allocating Julia functions allocate an output buffer and then call the non-allocating method.

All-to-all collective communications support in place operations by passing MPI.IN_PLACE with the same syntax documented by MPI. One-to-All communications support it by calling the function *_in_place!, calls the MPI functions with the right syntax on root and non root process.

MPI.Allgather! — Function.

Allgather!(sendbuf, recvbuf, count, comm)

Each process sends the first count elements of sendbuf to the other processes, who store the results in rank order into recvbuf.

If sendbuf==MPI.IN_PLACE the input data is assumed to be in the area of recvbuf where the process would receive it's own contribution.

source

Allgather!(buf, count, comm)

Equivalent to Allgather!(MPI.IN_PLACE, buf, count, comm).

source

MPI.Allgather — Function.

Allgather(sendbuf, count, comm)

Each process sends the first count elements of sendbuf to the other processes, who store the results in rank order allocating the output buffer.

source

MPI.Allgatherv! — Function.

Allgatherv!(sendbuf, recvbuf, counts, comm)

Each process sends the first counts[rank] elements of the buffer sendbuf to all other process. Each process stores the received data in rank order in the buffer recvbuf.

if sendbuf==MPI.IN_PLACE every process takes the data to be sent is taken from the interval of recvbuf where it would store it's own data.

source

MPI.Allgatherv — Function.

Allgatherv(sendbuf, counts, comm)

Each process sends the first counts[rank] elements of the buffer sendbuf to all other process. Each process allocates an output buffer and stores the received data in rank order.

source

MPI.Allreduce! — Function.

Allreduce!(sendbuf, recvbuf[, count=length(sendbuf)], op, comm)

Performs op reduction on the first count elements of the buffer sendbuf storing the result in the recvbuf of all processes in the group.

All-reduce is equivalent to a Reduce! operation followed by a Bcast!, but can lead to better performance.

If sendbuf==MPI.IN_PLACE the data is read from recvbuf and then overwritten with the results.

To handle allocation of the output buffer, see Allreduce.

source

Allreduce!(buf, op, comm)

Performs op reduction in place on the buffer sendbuf, overwriting it with the results on all the processes in the group.

Equivalent to calling Allreduce!(MPI.IN_PLACE, buf, op, comm)

source

MPI.Allreduce — Function.

Allreduce(sendbuf, op, comm)

Performs op reduction on the buffer sendbuf, allocating and returning the output buffer in all processes of the group.

To specify the output buffer or perform the operation in pace, see Allreduce!.

source

MPI.Alltoall! — Function.

Alltoall!(sendbuf, recvbuf, count, comm)

Every process divides the buffer sendbuf into Comm_size(comm) chunks of length count, sending the j-th chunk to the j-th process. Every process stores the data received from the j-th process in the j-th chunk of the buffer recvbuf.

rank    send buf                        recv buf
----    --------                        --------
 0      a,b,c,d,e,f       Alltoall      a,b,A,B,α,β
 1      A,B,C,D,E,F  ---------------->  c,d,C,D,γ,ψ
 2      α,β,γ,ψ,η,ν                     e,f,E,F,η,ν

If sendbuf==MPI.IN_PLACE, data is sent from the recvbuf and then overwritten.

source

MPI.Alltoall — Function.

Alltoall(sendbuf, count, comm)

Every process divides the buffer sendbuf into Comm_size(comm) chunks of length count, sending the j-th chunk to the j-th process. Every process allocates the output buffer and stores the data received from the j-th process in the j-th chunk.

rank    send buf                        recv buf
----    --------                        --------
 0      a,b,c,d,e,f       Alltoall      a,b,A,B,α,β
 1      A,B,C,D,E,F  ---------------->  c,d,C,D,γ,ψ
 2      α,β,γ,ψ,η,ν                     e,f,E,F,η,ν

source

MPI.Alltoallv! — Function.

Alltoallv!(sendbuf::T, recvbuf::T, scounts, rcounts, comm)

MPI.IN_PLACE is not supported for this operation.

source

Missing docstring.

Missing docstring for MPI.Alltoallv. Check Documenter's build log for details.

MPI.Barrier — Function.

Barrier(comm::Comm)

Blocks until comm is synchronized.

If comm is an intracommunicator, then it blocks until all members of the group have called it.

If comm is an intercommunicator, then it blocks until all members of the other group have called it.

source

MPI.Bcast! — Function.

Bcast!(buf[, count=length(buf)], root, comm::Comm)

Broadcast the first count elements of the buffer buf from root to all processes.

source

Missing docstring.

Missing docstring for MPI.Bcast!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Exscan. Check Documenter's build log for details.

MPI.Gather! — Function.

Gather!(sendbuf, recvbuf, count, root, comm)

Each process sends the first count elements of the buffer sendbuf to the root process. The root process stores elements in rank order in the buffer buffer recvbuf.

count should be the same for all processes. If the number of elements varies between processes, use Gatherv! instead.

To perform the reduction in place refer to Gather_in_place!.

source

MPI.Gather — Function.

Gather(sendbuf[, count=length(sendbuf)], root, comm)

Each process sends the first count elements of the buffer sendbuf to the root process. The root allocates the output buffer and stores elements in rank order.

source

MPI.Gather_in_place! — Function.

Gather_in_place!(buf, count, root, comm)

Each process sends the first count elements of the buffer buf to the root process. The root process stores elements in rank order in the buffer buffer buf, sending no data to itself.

This is functionally equivalent to calling

if root == MPI.Comm_rank(comm)
    Gather!(MPI.IN_PLACE, buf, count, root, comm)
else
    Gather!(buf, C_NULL, count, root, comm)
end

source

MPI.Gatherv! — Function.

Gatherv!(sendbuf, recvbuf, counts, root, comm)

Each process sends the first counts[rank] elements of the buffer sendbuf to the root process. The root stores elements in rank order in the buffer recvbuf.

To perform the reduction in place refer to Gatherv_in_place!.

source

MPI.Gatherv — Function.

Gatherv(sendbuf, counts, root, comm)

Each process sends the first counts[rank] elements of the buffer sendbuf to the root process. The root allocates the output buffer and stores elements in rank order.

source

MPI.Gatherv_in_place! — Function.

Gatherv_in_place!(buf, counts, root, comm)

Each process sends the first counts[rank] elements of the buffer buf to the root process. The root allocates the output buffer and stores elements in rank order.

This is functionally equivalent to calling

if root == MPI.Comm_rank(comm)
    Gatherv!(MPI.IN_PLACE, buf, counts, root, comm)
else
    Gatherv!(buf, C_NULL, counts, root, comm)
end

source

MPI.Reduce! — Function.

Reduce!(sendbuf, recvbuf[, count=length(sendbuf)], op, root, comm)

Performs op reduction on the first count elements of the buffer sendbuf and stores the result in recvbuf on the process of rank root.

On non-root processes recvbuf is ignored.

To perform the reduction in place, see Reduce_in_place!.

To handle allocation of the output buffer, see Reduce.

source

MPI.Reduce — Function.

Reduce(sendbuf, count, op, root, comm)

Performs op reduction on the buffer sendbuf and stores the result in an output buffer allocated on the process of rank root. An empty array will be returned on all other processes.

To specify the output buffer, see Reduce!.

To perform the reduction in place, see Reduce_in_place!.

source

MPI.Reduce_in_place! — Function.

Reduce_in_place!(buf, count, op, root, comm)

Performs op reduction on the first count elements of the buffer buf and stores the result on buf of the root process in the group.

This is equivalent to calling

if root == MPI.Comm_rank(comm)
    Reduce!(MPI.IN_PLACE, buf, count, root, comm)
else
    Reduce!(buf, C_NULL, count, root, comm)
end

To handle allocation of the output buffer, see Reduce.

To specify a separate output buffer, see Reduce!.

source

Missing docstring.

Missing docstring for MPI.Scan. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Scan. Check Documenter's build log for details.

MPI.Scatter! — Function.

Scatter!(sendbuf, recvbuf, count, root, comm)

Splits the buffer sendbuf in the root process into Comm_size(comm) chunks and sends the j-th chunk to the process of rank j into the recvbuf buffer, which must be of length at least count.

count should be the same for all processes. If the number of elements varies between processes, use Scatter! instead.

To perform the reduction in place, see Scatter_in_place!.

To handle allocation of the output buffer, see Scatter.

source

MPI.Scatter — Function.

Scatter(sendbuf, count, root, comm)

Splits the buffer sendbuf in the root process into Comm_size(comm) chunks and sends the j-th chunk to the process of rank j, allocating the output buffer.

source

MPI.Scatter_in_place! — Function.

Scatter_in_place!(buf, count, root, comm)

Splits the buffer buf in the root process into Comm_size(comm) chunks and sends the j-th chunk to the process of rank j. No data is sent to the root process.

This is functionally equivalent to calling

if root == MPI.Comm_rank(comm)
    Scatter!(buf, MPI.IN_PLACE, count, root, comm)
else
    Scatter!(C_NULL, buf, count, root, comm)
end

To specify a separate output buffer, see Scatter!.

To handle allocation of the output buffer, see Scatter.

source

MPI.Scatterv! — Function.

Scatterv!(sendbuf, recvbuf, counts, root, comm)

Splits the buffer sendbuf in the root process into Comm_size(comm) chunks of length counts[j] and sends the j-th chunk to the process of rank j into the recvbuf buffer, which must be of length at least count.

To perform the reduction in place refer to Scatterv_in_place!.

source

MPI.Scatterv — Function.

Scatterv(sendbuf, counts, root, comm)

Splits the buffer sendbuf in the root process into Comm_size(comm) chunks of length counts[j] and sends the j-th chunk to the process of rank j, which allocates the output buffer

source

MPI.Scatterv_in_place! — Function.

Scatterv_in_place(buf, counts, root, comm)

Splits the buffer buf in the root process into Comm_size(comm) chunks of length counts[j] and sends the j-th chunk to the process of rank j into the buf buffer, which must be of length at least count. The root process sends nothing to itself.

This is functionally equivalent to calling

if root == MPI.Comm_rank(comm)
    Scatterv!(buf, MPI.IN_PLACE, counts, root, comm)
else
    Scatterv!(C_NULL, buf, counts, root, comm)
end

source

One-sided communication

Julia Function (assuming import MPI)	Fortran Function

MPI.Win_create	MPI_Win_create

MPI.Win_create_dynamic	MPI_Win_create_dynamic

MPI.Win_allocate_shared	MPI_Win_allocate_shared

MPI.Win_shared_query	MPI_Win_shared_query

MPI.Win_attach	MPI_Win_attach

MPI.Win_detach	MPI_Win_detach

MPI.Win_fence	MPI_Win_fence

MPI.Win_flush	MPI_Win_flush

MPI.Win_free	MPI_Win_free

MPI.Win_sync	MPI_Win_sync

MPI.Win_lock	MPI_Win_lock

MPI.Win_unlock	MPI_Win_unlock

MPI.Get	MPI_Get

MPI.Put	MPI_Put

MPI.Fetch_and_op	MPI_Fetch_and_op

MPI.Accumulate	MPI_Accumulate

MPI.Get_accumulate	MPI_Get_accumulate

Missing docstring.

Missing docstring for MPI.Win_create. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_create_dynamic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_allocate_shared. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_shared_query. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_attach. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_detach. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_fence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_flush. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_free. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_sync. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_lock. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Win_unlock. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Get. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Put. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Fetch_and_op. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Accumulate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPI.Get_accumulate. Check Documenter's build log for details.