Library information

Library information

Constants

MPI.MPI_VERSION — Constant.

MPI_VERSION :: VersionNumber

The supported version of the MPI standard.

External links

MPI_Get_version man page: OpenMPI, MPICH

source

MPI.MPI_LIBRARY — Constant.

MPI_LIBRARY :: MPIImpl

The current MPI implementation: this is determined by

See also

MPIImpl

source

MPI.MPI_LIBRARY_VERSION — Constant.

MPI_LIBRARY_VERSION :: VersionNumber

The version of the MPI library

source

MPI.MPI_LIBRARY_ABI — Constant.

MPI_LIBRARY_ABI :: MPIABI

The ABI used by the current MPI implementation.

See also

MPIABI

source

MPI.MPI_LIBRARY_VERSION_STRING — Constant.

MPI_LIBRARY_VERSION_STRING :: String

The full version string provided by the library

External links

MPI_Get_library_version man page: OpenMPI, MPICH

source

Enums

MPI.MPIImpl — Type.

MPIImpl

An enum corresponding to known MPI implementations

UnknownMPI: unable to determine MPI implementation

MPICH: MPICH

OpenMPI: Open MPI

MicrosoftMPI: Microsoft MPI

IntelMPI: Intel MPI

SpectrimMPI: IBM Spectrum MPI

MVAPICH: MVAPICH

See also

MPI_LIBRARY

source

MPI.MPIABI — Type.

MPIABI

An enum corresponding to known MPI Application Binary Interfaces (ABI)

UnknownABI: unable to determine MPI ABI. MPI.jl will attempt to build a small C program to determine the necessary constants and type information.

MPICHABI: Compatible with MPICH ABI Compatibility Initiative.

OpenMPIABI: Compatible with Open MPI.

MicrosftMPIABI: Compatible with Microsoft MPI.

source

Functions

MPI.has_cuda — Function.

MPI.has_cuda()

Check if the MPI implementation is known to have CUDA support. Currently only Open MPI provides a mechanism to check, so it will return false with other implementations (unless overriden).

This can be overriden by setting the JULIA_MPI_HAS_CUDA environment variable to true or false.

source

MPI.identify_implementation — Function.

impl, version, abi = identify_implementation()

Attempt to identify the MPI implementation based on MPI_LIBRARY_VERSION_STRING. Returns a triple of values:

impl: a value of type MPIImpl

version: a VersionNumber of the library, or nothing if it cannot be determined.

abi: a value of MPIABI. This can be overridden by the JULIA_MPI_ABI environment variable.

This function is only intended for internal use. Users should use MPI_LIBRARY, MPI_LIBRARY_VERSION or MPI_LIBRARY_ABI.

source