Topology

MPI.Dims_create — Function

newdims = Dims_create(nnodes::Integer, dims)

A convenience function for selecting a balanced Cartesian grid of a total of nnodes nodes, for example to use with MPI.Cart_create.

dims is an array or tuple of integers specifying the number of nodes in each dimension. The function returns an array newdims of the same length, such that if newdims[i] = dims[i] if dims[i] is non-zero, and prod(newdims) == nnodes, and values newdims are as close to each other as possible.

nnodes should be divisible by the product of the non-zero entries of dims.

External links

MPI_Dims_create man page: OpenMPI, MPICH

MPI.Cart_create — Function

comm_cart = Cart_create(comm::Comm, dims; periodic=map(_->false, dims), reorder=false)

Create new MPI communicator with Cartesian topology information attached.

dims is an array or tuple of integers specifying the number of MPI processes in each coordinate direction, and periodic is an array or tuple of Bools indicating the periodicity of each coordinate. prod(dims) must be less than or equal to the size of comm; if it is smaller than some processes are returned a null communicator.

If reorder == false then the rank of each process in the new group is identical to its rank in the old group, otherwise the function may reorder the processes.

See also MPI.Dims_create.

External links

MPI_Cart_create man page: OpenMPI, MPICH

MPI.Cart_get — Function

dims, periods, coords = Cart_get(comm::Comm)

Obtain information on the Cartesian topology of dimension N underlying the communicator comm. This is specified by two Cint arrays of N elements for the number of processes and periodicity properties along each Cartesian dimension. A third Cint array is returned, containing the Cartesian coordinates of the calling process.

External links

MPI_Cart_get man page: OpenMPI, MPICH

MPI.Cart_coords — Function

coords = Cart_coords(comm::Comm, rank::Integer=Comm_rank(comm))

Determine coordinates of a process with rank rank in the Cartesian communicator comm. If no rank is provided, it returns the coordinates of the current process.

Returns an integer array of the 0-based coordinates. The inverse of Cart_rank.

External links

MPI_Cart_coords man page: OpenMPI, MPICH

MPI.Cart_rank — Function

rank = Cart_rank(comm::Comm, coords)

Determine process rank in communicator comm with Cartesian structure. The coords array specifies the 0-based Cartesian coordinates of the process. This is the inverse of MPI.Cart_coords

External links

MPI_Cart_rank man page: OpenMPI, MPICH

MPI.Cart_shift — Function

rank_source, rank_dest = Cart_shift(comm::Comm, direction::Integer, disp::Integer)

Return the source and destination ranks associated to a shift along a given direction.

External links

MPI_Cart_shift man page: OpenMPI, MPICH

MPI.Cart_sub — Function

comm_sub = Cart_sub(comm::Comm, remain_dims)

Create lower-dimensional Cartesian communicator from existent Cartesian topology.

remain_dims should be a boolean vector specifying the dimensions that should be kept in the generated subgrid.

External links

MPI_Cart_sub man page: OpenMPI, MPICH

MPI.Cartdim_get — Function

ndims = Cartdim_get(comm::Comm)

Return number of dimensions of the Cartesian topology associated with the communicator comm.

External links

MPI_Cartdim_get man page: OpenMPI, MPICH