Topology

MPI.Cartdim_get — Function

Cartdim_get(comm::Comm)

Return number of dimensions of the Cartesian topology associated with the communicator comm.

MPI.Cart_get — Function

Cart_get(comm::Comm, maxdims::Integer)

Obtain information on the Cartesian topology of dimension maxdims underlying the communicator comm. This is specified by two Cint arrays of maxdims elements for the number of processes and periodicity properties along each Cartesian dimension. A third Cint array is returned, containing the Cartesian coordinates of the calling process.